requirements.txt
# Django and REST Framework
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0,<4.0.0
django-cors-headers>=4.0.0,<5.0.0
django-filter>=23.0,<24.0
djangorestframework-simplejwt>=5.0.0,<6.0.0

# Database
psycopg2-binary>=2.9.5,<3.0.0
pgvector>=0.2.0,<1.0.0

# OpenAI and utilities
openai>=1.0.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
requests>=2.28.0,<3.0.0

# File processing
python-magic>=0.4.27,<1.0.0
PyPDF2>=3.0.0,<4.0.0

# Utilities
Pillow>=9.4.0,<10.0.0
tqdm>=4.65.0,<5.0.0
numpy>=1.24.0,<2.0.0

# Development tools
pytest>=7.0.0,<8.0.0
pytest-django>=4.5.0,<5.0.0
pytest-mock>=3.10.0,<4.0.0
black>=23.0.0,<24.0.0
flake8>=6.0.0,<7.0.0 